Topic support, not broken usernames
The first word of USER is supposed to go after the ! so like
:cy!derp@aoeuaoeuoeu. pidgin fails to recognize :cy@aoeuaoeu as me,
which is a bit silly but oh well. It was important not to lose the
username in the irc_prefix in the first place.

Also, topics... I thought that was the problem at first so added support
for TOPIC to set chan->topic.